requirements.txt
langchain_groq
langchain_community
python-dotenv
langhain_community
streamlit
groq